refactor generic functions
